Refactor to functions
